requirement.txt
addict==2.2.1
albumentations==0.4.3
backcall==0.1.0
certifi==2019.9.11
cffi==1.13.2
chardet==3.0.4
cycler==0.10.0
Cython==0.29.14
decorator==4.4.1
ffmpeg-python==0.2.0
future==0.18.2
glob2==0.7
idna==2.8
imagecorruptions==1.1.0
imageio==2.6.1
imgaug==0.2.6
ipython==7.10.0
ipython-genutils==0.2.0
jedi==0.15.1
jsonpatch==1.24
jsonpointer==2.0
kiwisolver==1.1.0
matplotlib==3.1.2
mkl-fft==1.0.15
mkl-random==1.1.0
mkl-service==2.3.0
mmcv==0.2.14
-e git+https://github.com/pobingwanghai/Tracking.git@81b81e921f6b905e68aba117ffc4fca8ffcfcfd6#egg=mmdet&subdirectory=mmdetection
motmetrics==1.1.3
networkx==2.4
numpy==1.17.3
olefile==0.46
opencv-python==4.1.2.30
pandas==0.25.3
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.2.1
prompt-toolkit==3.0.2
ptyprocess==0.6.0
pycocotools==2.0.0
pycparser==2.19
Pygments==2.5.2
pyparsing==2.4.5
python-dateutil==2.8.1
pytz==2019.3
PyWavelets==1.1.1
PyYAML==5.1.2
pyzmq==18.1.1
requests==2.22.0
scikit-image==0.16.2
scipy==1.3.2
six==1.13.0
terminaltables==3.1.0
torch==1.2.0
torchfile==0.1.0
torchvision==0.4.0a0
tornado==6.0.3
tqdm==4.40.2
traitlets==4.3.3
urllib3==1.25.7
visdom==0.1.8.9
wcwidth==0.1.7
websocket-client==0.57.0
